Support tag types with conditional property-to-attribute maps. input, with "checked" is currently supported, other's can be added easily now.
Also, change so that attribute access on an AdvancedTag, if it does not find the requested property, will return "None" instead of raising AttributeError. This more closely follows JS syntax, especially for things like "checked" which may or may not be supported depending on the tag name.

Also, rename the "indent" property to _indent.