Add some useful methods to Tags, especially relating to blocks
- Update removeText to only remove the FIRST occurace of text (inline
with other javascript DOM functions). For old behaviour, see
removeTextAll which will remove all occurances of the text.

^^^ ^^^  ^^^   ^^^^ ^^^^ ^^
**MAYBE BACKWARDS INCOMPATIBLE** - depending on usage

- Add removeTextAll function to remove all occuracnes of text from text
nodes on an element.

- Change removeText to return the old block (text in that node prior to
replace). removeTextAll returns a list of all old blocks.

- Add addBlock / addBlocks functions which take generic blocks (may be a
str, may be an AdvancedTag ). Returns the added block.

- Add removeBlock / removeBlocks functions which take generic blocks
(may be str, may be AdvancedTag). Returns the removed tag or the block
of text prior to remove. None on removeBlock if none found, or None in
the corrosponding element in the list return of removeBlocks

- Add removeChildren function as a helper to remove multiple children.
  Returns the children removed in a list, with a "None" if that child
  was not present.